+++ src/hex/net/serializer.rs
@@ -43,9 +43,9 @@ impl DataSerializer<HexGame> for HexSerializer {
             .into_iter()
             .collect_vec();
 
-        let mut probs_vec = vec![0.0; (BOARD_SIZE * BOARD_SIZE) as usize];
+        let mut probs_vec = vec![0.0; BOARD_SIZE * BOARD_SIZE];
         for (m, prob) in probs {
-            probs_vec[m.to_idx() as usize] = prob;
+            probs_vec[m.to_idx()] = prob;
         }
 
         let json_obj = json::object! {
